Makefile.in - added vsnprintf configure.in - added vsnprintf magic gss-utils.c - series of improvements after code review gss-utils.h - improvements after code review remctl.c - changed type of file descriptors remctld.c - changed type of file descriptors
config.h.in - common header, for use with autoheader

+++ remctl.c
@@ -33,8 +33,8 @@ int use_syslog;    /* Toggle for sysctl vs stdout/stderr. */
 
 /* These are for storing either the socket for communication with client
    or the streams that talk to the network in case of inetd/tcpserver. */
-unsigned short READFD;
-unsigned short WRITEFD;
+int READFD;
+int WRITEFD;
 
 /* Used in establishing context with gss server. */
 gss_buffer_desc empty_token_buf = { 0, (void *)"" };
